fixed #20

+++ src/main/java/com/shelflife/project/controller/UserController.java
@@ -46,7 +46,7 @@ public class UserController {
 
         Optional<User> user = repo.findById(id);
 
-        if(!user.isPresent())
+        if (!user.isPresent())
             return ResponseEntity.badRequest().body(Map.of("error", "Invalid id"));
 
         return ResponseEntity.ok(user.get());
@@ -54,19 +54,22 @@ public class UserController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<?> deleteUser(@PathVariable long id, Authentication auth) {
-        if(auth == null || !auth.isAuthenticated())
+        if (auth == null || !auth.isAuthenticated())
             return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
-        
-        if (!repo.existsById(id))
-            return ResponseEntity.badRequest().body(Map.of("error", "Invalid id"));
 
         String userEmail = auth.getName();
         User self = repo.findByEmail(userEmail).get();
 
-        if(self.getId() == id)
+        if(!self.isAdmin())
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
+        if (!repo.existsById(id))
+            return ResponseEntity.badRequest().body(Map.of("error", "Invalid id"));
+
+        if (self.getId() == id)
             return ResponseEntity.status(HttpStatus.FORBIDDEN).body(Map.of("error", "You can't delete yourself"));
 
         repo.deleteById(id);
-        return ResponseEntity.ok(null);
+        return ResponseEntity.ok().build();
     }
 }
